tabled/ rename tests/matrix into tests/util
