+++ tabled/tests/core/table_test.rs
@@ -11,7 +11,7 @@ use tabled::{
     Table,
 };
 
-use crate::matrix::Matrix;
+use crate::util::Matrix;
 
 mod default_types {
     use super::*;
