+++ tabled/tests/main.rs
@@ -3,5 +3,4 @@ mod derive;
 mod macros;
 mod settings;
 
-#[cfg(feature = "std")]
-mod matrix;
+mod util;
