+++ tabled/tests/settings/disable_test.rs
@@ -11,7 +11,7 @@ use tabled::{
     },
 };
 
-use crate::matrix::Matrix;
+use crate::util::Matrix;
 
 test_table!(
     disable_rows,
