+++ tabled/tests/settings/duplicate_test.rs
@@ -9,7 +9,7 @@ use tabled::{
     },
 };
 
-use crate::matrix::Matrix;
+use crate::util::Matrix;
 
 test_table!(
     dup_cell_to_cell,
