+++ tabled/tests/settings/extract_test.rs
@@ -10,7 +10,7 @@ use tabled::{
     },
 };
 
-use crate::matrix::Matrix;
+use crate::util::Matrix;
 
 test_table!(
     extract_segment_full_test,
