+++ tabled/tests/settings/format_test.rs
@@ -9,7 +9,7 @@ use tabled::{
     },
 };
 
-use crate::matrix::Matrix;
+use crate::util::Matrix;
 
 #[cfg(feature = "ansi")]
 use tabled::settings::Color;
