+++ tabled/tests/settings/formatting_test.rs
@@ -7,7 +7,7 @@ use tabled::{
     Table,
 };
 
-use crate::matrix::Matrix;
+use crate::util::Matrix;
 
 test_table!(
     justification,
