+++ tabled/tests/settings/layout_test.rs
@@ -6,7 +6,7 @@ use tabled::{
     settings::{themes::Layout, Alignment, Style},
 };
 
-use crate::matrix::Matrix;
+use crate::util::Matrix;
 
 test_table!(
     theme_stick_left,
