+++ tabled/tests/settings/object_test.rs
@@ -10,7 +10,7 @@ use tabled::{
     },
 };
 
-use crate::matrix::Matrix;
+use crate::util::Matrix;
 
 // todo: Columns::all()
