+++ tabled/tests/settings/style_test.rs
@@ -16,7 +16,7 @@ use tabled::{
     Table,
 };
 
-use crate::matrix::Matrix;
+use crate::util::Matrix;
 
 test_table!(
     default_style,
