+++ tabled/tests/util/matrix.rs
