+++ tabled/tests/util/mod.rs
@@ -1,8 +1,11 @@
 #![allow(unused_imports)]
 
-#[allow(clippy::module_inception)]
+#[cfg(feature = "std")]
 mod matrix;
+#[cfg(feature = "std")]
 mod matrix_list;
 
+#[cfg(feature = "std")]
 pub use matrix::Matrix;
+#[cfg(feature = "std")]
 pub use matrix_list::MatrixList;
